refactor(ui): extract feed/inspector.ts — ContextInspectorPanel + 3 pure query helpers
